% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derive4dParam.R
\name{derive4dParam}
\alias{derive4dParam}
\title{extract and calculates Meteo Parameter from an ARPS netcdf file}
\source{
\tabular{ll}{
Package: \tab peRfectpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\usage{
derive4dParam(filename,param=c("tc","td","es","e","rh","pr","u","v","w","ws","wd"))
}
\arguments{
\item{filename}{is a filname of an ARPS netcdf file}

\item{param}{is a parameter(list)  out of "tc","td","es","e","rh","pr","u","v","w","ws","wd"}
}
\value{
derive4dParam returns the following parameters:
\tabular{ll}{
$paramter  \tab"tc","td","es","e","rh","pr","u","v","w","ws","wd"\cr
}
}
\description{
Extract arbitrary ARPS model run parameter for usage in R
}
\examples{
#### Example to extract thermodynamic or wind data
 #### from a ARPS nccdf file

arps.ncfile=system.file("kili.nc", package="aRps")
ws=derive4dParam(arps.ncfile,"ws")
ws
}
\author{
Chris Reudenbach and Hanna Meyer
}
\references{
\url{http://giswerk.org/doku.php?id=wac:modeling:arps:intro}
}
\seealso{
For calculate thermodynamic variables see\code{\link{calcMeteoParam}}.
}

