% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRefInfo.R
\name{getRefInfo}
\alias{getRefInfo}
\title{Calculate projection and extent from an ARPS netcdf file}
\source{
\tabular{ll}{
Package: \tab peRfectpeak\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\usage{
getRefInfo(file)
}
\arguments{
\item{file}{is a filname of an ARPS netcdf file}
}
\value{
getRefInfo returns the following parameters:
\tabular{ll}{
$proj \tab proj4 projection string\cr
$lat \tab central latitude of the data domain\cr
$lon \tab central longitude of the data domain\cr
$dx \tab grid spacing in x direction\cr
$dy \tab grid spacing in y direction\cr
$ext \tab xmin,xmax,ymin,ymax in $proj\cr
$coordx \tab list of x coordinates according to the domain $xcoord\cr
$coordy \tab list of y coordinates according to the domain $ycoord\cr
}
}
\description{
The information of the projection is calculated using by analysing the ARPS netcdf file parameters. Currently only "no_mapping" (default) and "lambert_conformal_conic" (north hemisphere) is supported.
\code{\link{getRefInfo}} provides the missing projection values of an ARPS netCDF file
}
\examples{
#### Example to extract the projection and
 #### calculate the extent coordinates from a ARPS nccdf file

arps.ncfile=system.file("kili.nc", package="aRps")
refInfo=getRefInfo(arps.ncfile)
refInfo
}
\author{
Chris Reudenbach and Hanna Meyer
}
\references{
\url{http://giswerk.org/doku.php?id=wac:modeling:arps:intro}
}
\seealso{
For retrieving variables see\code{\link{derive4dParam}}.

For calculate thermodynamic variables see\code{\link{calcMeteoParam}}.
}

