<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makenc3D.R
=======
>>>>>>> 587445933251a6d77ef860b5b8fe5d363fbd4f2b
\name{makenc3D}
\alias{makenc3D}
\title{Extract/calculate and georeferences meteo params from an ARPS netcdf file}
\source{
<<<<<<< HEAD
\tabular{ll}{
Package: \tab aRps\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
=======
\tabular{ll}{ Package: \tab aRps\cr Type: \tab Package\cr
Version: \tab 0.2\cr License: \tab GPL (>= 2)\cr LazyLoad:
\tab yes\cr }
>>>>>>> 587445933251a6d77ef860b5b8fe5d363fbd4f2b
}
\usage{
makenc3D(file,param)
}
\arguments{
<<<<<<< HEAD
\item{file}{is a filname of an ARPS netcdf file}

\item{param}{is name of the parameter to be extracted. if one use 'newAll' all params will be transcribed}
}
\value{
makenc3D returns the following parameters:
\tabular{ll}{
=======
  \item{file}{is a filname of an ARPS netcdf file}

  \item{param}{is name of the parameter to be extracted. if
  one use 'newAll' all params will be transcribed}
}
\value{
makenc3D returns the following parameters: \tabular{ll}{
>>>>>>> 587445933251a6d77ef860b5b8fe5d363fbd4f2b
outfile \tab new netcdf file with all or a single param\cr
}
}
\description{
<<<<<<< HEAD
The information of the projection is calculated using the informations as derived by the gdalinfo call of the ARPS netcdf file. Currently only "lambert_conformal_conic" and 'latlong' is supported.
\code{\link{getRefInfo}} provides all missing projection values of an ARPS netCDF file
=======
The information of the projection is calculated using the
informations as derived by the gdalinfo call of the ARPS
netcdf file. Currently only "lambert_conformal_conic" and
'latlong' is supported. \code{\link{getRefInfo}} provides
all missing projection values of an ARPS netCDF file
>>>>>>> 587445933251a6d77ef860b5b8fe5d363fbd4f2b
}
\examples{
#### Example to georeference an ARPS netCDF 3.0 file and to calculate some important meteo params

arpsexample=system.file("allgaeu_d1_ARP.nc", package="aRps")
makenc3D(arpsexample,'allNew')
}
\author{
Chris Reudenbach and Hanna MeyerQ
}
\references{
\url{http://giswerk.org/doku.php?id=doku:modeling:arps:arps_installation}
}
\seealso{
<<<<<<< HEAD
For retrieving and writing projected data see\code{\link{derive4dParam}}.
=======
For retrieving and writing projected data
see\code{\link{derive4dParam}}.
>>>>>>> 587445933251a6d77ef860b5b8fe5d363fbd4f2b
}

