% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotARPSskewT.R
\name{plotARPSskewT}
\alias{plotARPSskewT}
\title{Extract single ARPS columns and plots data as skew-T, log p diagram}
\source{
\tabular{ll}{
Package: \tab aRps\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\usage{
plotARPSskewT(file,col, row, tim, zoom = TRUE, winds=FALSE, writesounding=FALSE)
}
\arguments{
\item{file}{ARPS netCDF file}

\item{col}{array position of the column to be extracted}

\item{row}{array position of the row to be extracted}

\item{tim}{time step of the dataset to be extracted}

\item{zoom}{if (zoom=TRUE) the skewt log digramm is plotted only for a pressure between 1050 and 400 hPa}

\item{winds}{if (wind=TRUE) a windbarb column is added beside the digramm}

\item{writesounding}{if (writesounding=TRUE) showing result table}
}
\value{
plots a skew-T, log p diagram from the required domain position
returns additionale the corresponding table of data
}
\description{
plotARPSskewT extract a unique column of the arps ARPS cube. The derived single
data column is converted to a data.frame as used bei the plotradiosonde function of the adapted Radiosonde package
}
\examples{
# get data
 arps.ncfile=system.file("kili.nc", package="aRps")
 nc <- nc_open(arps.ncfile)

 # plots skew-T, log p diagram domain position 10, 10 at timepos 2
 plotARPSskewT(nc,10,10,2,zoom=TRUE, winds=FALSE,writesounding=FALSE)

 ### plots skew-T, log p diagram domain position 10, 10 at timepos 2 and shows data table
 plotARPSskewT(nc,10,10,2, zoom=FALSE, winds=TRUE,writesounding=TRUE)
}
\author{
Chris Reudenbach, Hanna Meyer
}
\keyword{keywords}

