% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getARPSsounding.R
\name{getARPSsounding}
\alias{getARPSsounding}
\title{Extract single ARPS columns and plots data as skew-T, log p diagram}
\source{
\tabular{ll}{
Package: \tab aRps\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\usage{
getARPSsounding(nc,col,row,tim)
}
\arguments{
\item{col}{array position of the column to be extracted}

\item{row}{array position of the row to be extracted}

\item{tim}{time step of the dataset to be extracted}

\item{filename}{of an valid ARPS netCDF file}

\item{zoom}{if (zoom=TRUE) the skewt log digramm is plotted only for a pressure between 1050 and 400 hPa}

\item{winds}{if (wind=TRUE) a windbarb column is added beside the digramm}

\item{viewtable}{if (viewtable=TRUE) showing result table}
}
\value{
a dataframe containing the above variables
}
\description{
getARPSsounding extract a arbitrary column of the ARPS 4D data cube.
The derived data column is converted to a data.frame
and can be plotted as a skew-T, log p diagram using plotradiosonde
}
\examples{
###############
 # get data
 arps.ncfile=system.file("kili.nc", package="aRps")
 nc <- nc_open(arps.ncfile)

 getARPSsounding(nc,10,10,2)
}
\author{
Chris Reudenbach, Hanna Meyer
}
\keyword{keywords}

