% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makenc3D.R
\name{makenc3D}
\alias{makenc3D}
\title{Extract/calculate and georeferences meteo params from an ARPS netcdf file}
\source{
\tabular{ll}{
Package: \tab aRps\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\usage{
makenc3D(file,param)
}
\arguments{
\item{file}{is a filname of an ARPS netcdf file}

\item{param}{is name of the parameter to be extracted. if one use 'newAll' all params will be transcribed}
}
\value{
makenc3D returns the following parameters:
\tabular{ll}{
outfile \tab new netcdf file with all or a single param\cr
}
}
\description{
The information of the projection is calculated using the informations as derived by the gdalinfo call of the ARPS netcdf file. Currently only "lambert_conformal_conic" and 'latlong' is supported.
\code{\link{getRefInfo}} provides all missing projection values of an ARPS netCDF file
}
\examples{
#### Example to georeference an ARPS netCDF 3.0 file and to calculate some important meteo params

arps.ncfile=system.file("kili.nc", package="aRps")
makenc3D(arps.ncfile,'allNew')
}
\author{
Chris Reudenbach and Hanna MeyerQ
}
\references{
\url{http://giswerk.org/doku.php?id=doku:modeling:arps:arps_installation}
}
\seealso{
For retrieving and writing projected data see\code{\link{derive4dParam}}.
}

