% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refARPSnc.R
\name{refARPSnc}
\alias{refARPSnc}
\title{Extract/calculate and georeferences meteo params from an ARPS netcdf file}
\source{
\tabular{ll}{
Package: \tab aRps\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\usage{
refARPSnc(file)
}
\arguments{
\item{file}{ARPS netCDF file}
}
\value{
refARPSnc returns the correctly georeferenced and formated ARPS netCDF file and adds windspeed and -direction and some basic thermodynamic variables
}
\description{
The information of the projection is calculated using the informations as derived by the gdalinfo call of the ARPS netcdf file. Currently only "lambert_conformal_conic" and 'latlong' is supported.
\code{\link{getRefInfo}} provides all missing projection values of an ARPS netCDF file
}
\examples{
###############
 #### Example to georeference an ARPS netCDF 3.0 file and to calculate some important meteo params

arps.ncfile=system.file("kili.nc", package="aRps")
 nc <- nc_open(arps.ncfile)
refARPSnc(arps.ncfile)
}
\author{
Chris Reudenbach and Hanna Meyer
}
\references{
\url{http://giswerk.org/doku.php?id=doku:modeling:arps:arps_installation}
}
\seealso{
If you want to use this data in a GIS or otherwise georeferenced you need to extract the projection and domain extent according to the reference system that was used by ARPS \code{\link{getRefInfo}}
}

